@@ -1,55 +1,54 @@
 {
   "author": "aleksandr-kiliushin",
   "dependencies": {
-    "@emotion/react": "11.7.1",
-    "@emotion/styled": "11.6.0",
-    "@hookform/resolvers": "2.8.8",
-    "@mui/icons-material": "5.2.5",
-    "@mui/material": "5.2.5",
-    "@reduxjs/toolkit": "1.7.1",
-    "chart.js": "3.6.2",
-    "date-fns": "2.27.0",
-    "react": "17.0.2",
-    "react-dom": "17.0.2",
-    "react-hook-form": "7.22.2",
-    "react-redux": "7.2.6",
-    "react-router-dom": "6.2.1",
-    "react-use": "17.3.2",
+    "@emotion/react": "11.10.4",
+    "@emotion/styled": "11.10.4",
+    "@hookform/resolvers": "2.9.8",
+    "@mui/icons-material": "5.10.3",
+    "@mui/material": "5.10.5",
+    "@reduxjs/toolkit": "1.8.5",
+    "chart.js": "3.9.1",
+    "date-fns": "2.29.3",
+    "history": "5.3.0",
+    "react": "18.2.0",
+    "react-dom": "18.2.0",
+    "react-hook-form": "7.35.0",
+    "react-redux": "8.0.2",
+    "react-router-dom": "6.4.0",
+    "react-use": "17.4.0",
     "yup": "0.32.11"
   },
   "devDependencies": {
-    "@babel/core": "7.16.5",
-    "@babel/preset-env": "7.16.5",
-    "@babel/preset-react": "7.16.5",
-    "@babel/preset-typescript": "7.16.5",
-    "@emotion/babel-preset-css-prop": "11.2.0",
-    "@testing-library/jest-dom": "5.16.1",
-    "@testing-library/react": "12.1.2",
-    "@testing-library/user-event": "13.5.0",
-    "@trivago/prettier-plugin-sort-imports": "3.1.1",
-    "@types/faker": "5.5.3",
-    "@types/jest": "27.0.3",
-    "@types/react": "17.0.37",
-    "@types/react-dom": "17.0.11",
-    "@types/react-router-dom": "5.3.2",
-    "@typescript-eslint/eslint-plugin": "5.8.0",
-    "@typescript-eslint/parser": "5.8.0",
-    "babel-jest": "27.4.5",
-    "babel-loader": "8.2.3",
-    "cypress": "10.6.0",
-    "eslint": "8.5.0",
-    "eslint-plugin-react": "7.28.0",
-    "html-loader": "3.0.1",
+    "@babel/core": "7.19.1",
+    "@babel/preset-env": "7.19.1",
+    "@babel/preset-react": "7.18.6",
+    "@babel/preset-typescript": "7.18.6",
+    "@emotion/babel-preset-css-prop": "11.10.0",
+    "@testing-library/jest-dom": "5.16.5",
+    "@testing-library/react": "13.4.0",
+    "@testing-library/user-event": "14.4.3",
+    "@trivago/prettier-plugin-sort-imports": "3.3.0",
+    "@types/jest": "29.0.3",
+    "@types/react": "18.0.20",
+    "@types/react-dom": "18.0.6",
+    "@types/react-router-dom": "5.3.3",
+    "@typescript-eslint/eslint-plugin": "5.37.0",
+    "@typescript-eslint/parser": "5.37.0",
+    "babel-jest": "29.0.3",
+    "babel-loader": "8.2.5",
+    "cypress": "10.8.0",
+    "eslint": "8.23.1",
+    "eslint-plugin-react": "7.31.8",
     "html-webpack-plugin": "5.5.0",
-    "husky": "^8.0.1",
-    "jest": "27.4.5",
-    "lint-staged": "^13.0.3",
-    "style-loader": "3.3.1",
+    "husky": "8.0.1",
+    "jest": "29.0.3",
+    "jest-environment-jsdom": "29.0.3",
+    "lint-staged": "13.0.3",
     "ts-node": "10.9.1",
     "typescript": "4.8.3",
-    "webpack": "5.65.0",
-    "webpack-cli": "4.9.1",
-    "webpack-dev-server": "4.6.0",
+    "webpack": "5.74.0",
+    "webpack-cli": "4.10.0",
+    "webpack-dev-server": "4.11.0",
     "whatwg-fetch": "3.6.2"
   },
   "license": "ISC",
