@@ -1,7 +1,7 @@
 import { Global } from "@emotion/react"
 import ThemeProvider from "@mui/material/styles/ThemeProvider"
 import { StrictMode } from "react"
-import ReactDOM from "react-dom"
+import { createRoot } from "react-dom/client"
 import { Provider } from "react-redux"
 import { BrowserRouter } from "react-router-dom"
 
@@ -10,7 +10,16 @@ import { globalStyles } from "#styles/global"
 import { theme } from "#styles/theme"
 import { App } from "#views"
 
-ReactDOM.render(
+let container = document.querySelector("#root")
+if (container === null) {
+  container = document.createElement("div")
+  container.setAttribute("id", "root")
+  document.body.append(container)
+}
+
+const root = createRoot(container)
+
+root.render(
   <StrictMode>
     <BrowserRouter>
       <Provider store={store}>
@@ -20,6 +29,5 @@ ReactDOM.render(
         </ThemeProvider>
       </Provider>
     </BrowserRouter>
-  </StrictMode>,
-  document.querySelector("#root")
+  </StrictMode>
 )
