@@ -7,7 +7,7 @@ import { Auth } from "./index"
 
 describe("Auth service.", () => {
   test("Login works correctly.", async () => {
-    render(<Auth />, { iAm: "guest" })
+    await render(<Auth />, { iAm: "guest" })
 
     expect(localStorage.authToken).toBeUndefined()
     expect(screen.getByText("Welcome")).toBeInTheDocument()
@@ -15,7 +15,7 @@ describe("Auth service.", () => {
     await waitFor(() => userEvent.type(screen.getByLabelText("Username"), "john-doe"))
     await waitFor(() => userEvent.type(screen.getByLabelText("Password"), "john-doe-password"))
     expect(screen.getByText("Log in")).toBeEnabled()
-    userEvent.click(screen.getByText("Log in"))
+    await waitFor(() => userEvent.click(screen.getByText("Log in")))
     await waitForElementToBeRemoved(() => screen.getByText("Welcome"))
     expect(localStorage.authToken).toEqual(expect.stringMatching(".+"))
     expect(screen.queryByText("Log in")).not.toBeInTheDocument()
@@ -26,12 +26,12 @@ describe("Auth service.", () => {
   })
 
   test("Case: User enters unexisting username.", async () => {
-    render(<Auth />, { iAm: "guest" })
+    await render(<Auth />, { iAm: "guest" })
 
     expect(localStorage.authToken).toBeUndefined()
     await waitFor(() => userEvent.type(screen.getByLabelText("Username"), "john-doe-INCORRECT-USERNAME"))
     await waitFor(() => userEvent.type(screen.getByLabelText("Password"), "john-doe-password"))
-    userEvent.click(screen.getByText("Log in"))
+    await waitFor(() => userEvent.click(screen.getByText("Log in")))
     await waitFor(() => expect(screen.getByText("User not found.")).toBeInTheDocument())
 
     userEvent.clear(screen.getByLabelText("Username"))
@@ -44,7 +44,7 @@ describe("Auth service.", () => {
   })
 
   test("Case: User enters unexisting username.", async () => {
-    render(<Auth />, { iAm: "guest" })
+    await render(<Auth />, { iAm: "guest" })
 
     expect(localStorage.authToken).toBeUndefined()
     await waitFor(() => userEvent.type(screen.getByLabelText("Username"), "john-doe"))
@@ -66,7 +66,7 @@ describe("Auth service.", () => {
 
     expect(localStorage.authToken).toEqual(expect.stringMatching(".+"))
     userEvent.click(screen.getByText("Log out"))
-    expect(localStorage.authToken).toBeUndefined()
+    await waitFor(() => expect(localStorage.authToken).toBeUndefined())
     expect(screen.queryByText("Log out")).not.toBeInTheDocument()
     expect(screen.getByText("Log in")).toBeInTheDocument()
   })
