@@ -2,7 +2,6 @@ import { waitFor } from "@testing-library/react"
 
 import { getAuthToken } from "#utils/getAuthToken"
 import { render } from "#utils/testing/render"
-import { wait } from "#utils/wait"
 
 import { App } from "./index"
 
@@ -15,16 +14,14 @@ describe("<App />", () => {
 
   test("If an authorized user open some page, they are NOT redirected to /auth..", async () => {
     const { history } = await render(<App />, { iAm: "john-doe", initialUrl: "/settings" })
-    await wait(1000)
 
     expect(history.location.pathname).toEqual("/settings")
   })
 
-  test("If the user has authToken in localStorage, they is NOT redirected to /auth.", async () => {
+  test("If the user has authToken in localStorage (i. e. was logged in before and refreshes the page), they is NOT redirected to /auth.", async () => {
     localStorage.authToken = await getAuthToken("john-doe")
     const { history } = await render(<App />, { iAm: undefined, initialUrl: "/settings" })
 
-    await wait(1000)
     expect(history.location.pathname).toEqual("/settings")
   })
 })
