@@ -18,7 +18,7 @@ describe("Finance categories service", () => {
     await render(<Settings />, { iAm: "john-doe" })
 
     userEvent.click(screen.getByText("+New"))
-    expect(screen.getByText("Create category")).toBeInTheDocument()
+    expect(await screen.findByText("Create category")).toBeInTheDocument()
     expect(screen.getByText("Cancel")).toBeInTheDocument()
     expect(screen.getByText("Submit")).toBeInTheDocument()
   })
@@ -27,40 +27,39 @@ describe("Finance categories service", () => {
     await render(<Settings />, { iAm: "john-doe" })
 
     userEvent.click(screen.getByText("+New"))
-    expect(screen.getByText("Create category")).toBeInTheDocument()
+    expect(await screen.findByRole("dialog")).toBeInTheDocument()
     userEvent.click(screen.getByText("Cancel"))
-    expect(screen.queryByText("Create category")).not.toBeInTheDocument()
+    await waitForElementToBeRemoved(() => screen.getByRole("dialog"))
   })
 
   test("A new category is created correctly.", async () => {
     await render(<Settings />, { iAm: "john-doe" })
 
     userEvent.click(screen.getByText("+New"))
-    expect(screen.getByText("Submit")).toBeDisabled()
-    userEvent.type(screen.getByLabelText("Name"), "travel")
+    expect(await screen.findByText("Submit")).toBeDisabled()
+    await waitFor(() => userEvent.type(screen.getByLabelText("Name"), "travel"))
     await waitFor(() => userEvent.click(screen.getByLabelText("expense")))
-    expect(screen.getByText("Submit")).toBeEnabled()
+    await waitFor(() => expect(screen.getByText("Submit")).toBeEnabled())
     userEvent.click(screen.getByText("Submit"))
     await waitForElementToBeRemoved(() => screen.getByRole("dialog"))
-    expect(await screen.findByRole("cell", { name: "travel" })).toBeInTheDocument()
+    await waitFor(() => expect(screen.getByRole("cell", { name: "travel" })).toBeInTheDocument())
   })
 
   test("case: user tries to create a category that already exists and then fixes input values.", async () => {
     await render(<Settings />, { iAm: "john-doe" })
 
     userEvent.click(screen.getByText("+New"))
-    userEvent.type(screen.getByLabelText("Name"), "education")
+    await waitFor(() => userEvent.type(screen.getByLabelText("Name"), "education"))
     await waitFor(() => userEvent.click(screen.getByLabelText("expense")))
     userEvent.click(screen.getByText("Submit"))
     expect(await screen.findAllByText('"education" expense category already exists.')).toHaveLength(2)
     expect(screen.getByText("Submit")).toBeDisabled()
-    userEvent.clear(screen.getByLabelText("Name"))
+    await waitFor(() => userEvent.clear(screen.getByLabelText("Name")))
     await waitFor(() => expect(screen.getAllByText('"education" expense category already exists.')).toHaveLength(1))
-    userEvent.type(screen.getByLabelText("Name"), "teaching")
-    await waitFor(() => expect(screen.getByText("Submit")).toBeDisabled())
-    userEvent.click(screen.getByLabelText("income"))
+    await waitFor(() => userEvent.type(screen.getByLabelText("Name"), "teaching"))
+    await waitFor(() => userEvent.click(screen.getByLabelText("income")))
     await waitFor(() => expect(screen.queryAllByText('"education" expense category already exists.')).toHaveLength(0))
-    userEvent.click(screen.getByText("Submit"))
+    await waitFor(() => userEvent.click(screen.getByText("Submit")))
     await waitForElementToBeRemoved(() => screen.getByRole("dialog"))
     expect(await screen.findByRole("cell", { name: "teaching" })).toBeInTheDocument()
   })
@@ -108,7 +107,7 @@ describe("Finance categories service", () => {
 
     expect(await screen.findByText("salary")).toBeInTheDocument()
     userEvent.click(screen.getByTestId("salary-income-category-delete-button"))
-    expect(screen.getByRole("dialog")).toBeInTheDocument()
+    await waitFor(() => expect(screen.getByRole("dialog")).toBeInTheDocument())
     expect(screen.getByRole("heading", { name: "Delete category" })).toBeInTheDocument()
     userEvent.click(screen.getByText("Delete"))
     await waitForElementToBeRemoved(() => screen.getByRole("dialog"))
